Creates Errors.js for error and error checking.
